+++ module2.3.py
@@ -1,5 +1,5 @@
-rec_length = int(input("Enter length of the rectangle: "))
-rec_width = int(input("Enter width of the rectangle: "))
+rec_length = float(input("Enter length of the rectangle: "))
+rec_width = float(input("Enter width of the rectangle: "))
 
 rec_perimeter = (rec_length + rec_width) * 2
 print("The perimeter is", rec_perimeter)
